add TokenBlocklist in models.py
This could be expanded to fit the needs of your application. For example, it could track who revoked a JWT, when a token expires, notes for why a JWT was revoked, an endpoint to un-revoked a JWT, etc. Making jti an index can significantly speed up the search when there are tens of thousands of records. Remember this query will happen for every (protected) request, If your database supports a UUID type, this can be used for the jti column as well